+++ handler/api/users/create.go
@@ -35,7 +35,7 @@ type userWithToken struct {
 // to create the named user account in the system.
 func HandleCreate(users core.UserStore, service core.UserService, sender core.WebhookSender) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		in := new(core.User)
+		in := new(userWithToken)
 		err := json.NewDecoder(r.Body).Decode(in)
 		if err != nil {
 			render.BadRequest(w, err)