+++ handler/api/users/create_test.go
@@ -67,6 +67,58 @@ func TestCreate(t *testing.T) {
 	}
 }
 
+func TestCreateWithToken(t *testing.T) {
+	controller := gomock.NewController(t)
+	defer controller.Finish()
+
+	users := mock.NewMockUserStore(controller)
+	users.EXPECT().Create(gomock.Any(), gomock.Any()).Do(func(_ context.Context, in *core.User) error {
+		if got, want := in.Login, "octocat"; got != want {
+			t.Errorf("Want user login %s, got %s", want, got)
+		}
+		if got, want := in.Machine, true; got != want {
+			t.Errorf("Want user machine %v, got %v", want, got)
+		}
+		if got, want := in.Hash, "abc123"; got != want {
+			t.Errorf("Want user hash %s, got %s", want, got)
+		}
+		return nil
+	})
+
+	webhook := mock.NewMockWebhookSender(controller)
+	webhook.EXPECT().Send(gomock.Any(), gomock.Any()).Return(nil)
+
+	service := mock.NewMockUserService(controller)
+
+	in := new(bytes.Buffer)
+	json.NewEncoder(in).Encode(&userWithToken{&core.User{Login: "octocat", Machine: true}, "abc123"})
+	w := httptest.NewRecorder()
+	r := httptest.NewRequest("POST", "/", in)
+
+	HandleCreate(users, service, webhook)(w, r)
+	if got, want := w.Code, 200; want != got {
+		t.Errorf("Want response code %d, got %d", want, got)
+	}
+
+	out := new(userWithToken)
+	json.NewDecoder(w.Body).Decode(out)
+	if got, want := out.Login, "octocat"; got != want {
+		t.Errorf("Want user login %s, got %s", want, got)
+	}
+	if got, want := out.Active, true; got != want {
+		t.Errorf("Want user active %v, got %v", want, got)
+	}
+	if got := out.Created; got == 0 {
+		t.Errorf("Want user created set to current unix timestamp, got %v", got)
+	}
+	if got := out.Updated; got == 0 {
+		t.Errorf("Want user updated set to current unix timestamp, got %v", got)
+	}
+	if got, want := out.Token, "abc123"; got != want {
+		t.Errorf("Want user token %s, got %s", want, got)
+	}
+}
+
 func TestCreate_CorrectName(t *testing.T) {
 	controller := gomock.NewController(t)
 	defer controller.Finish()